Move into commands package, and make functions private
