+++ internal/commands/completion.go
@@ -51,7 +51,7 @@ To configure your zsh shell to load completions for each session, add the follow
 	. $(pack completion --shell zsh)
   
 	`,
-		RunE: LogError(logger, func(cmd *cobra.Command, args []string) error {
+		RunE: logError(logger, func(cmd *cobra.Command, args []string) error {
 			completionFunc, ok := shellExtensions[flags.Shell]
 			if !ok {
 				return errors.Errorf("%s is unsupported shell", flags.Shell)
