+++ internal/commands/inspect_builder.go
@@ -11,6 +11,10 @@ import (
 	"github.com/buildpacks/pack/logging"
 )
 
+type BuilderInspector interface {
+	InspectBuilder(name string, daemon bool, modifiers ...pack.BuilderInspectionModifier) (*pack.BuilderInfo, error)
+}
+
 type InspectBuilderFlags struct {
 	Depth        int
 	OutputFormat string
@@ -28,21 +32,21 @@ func InspectBuilder(
 		Args:    cobra.MaximumNArgs(2),
 		Short:   "Show information about a builder",
 		Example: "pack inspect-builder cnbs/sample-builder:bionic",
-		RunE: LogError(logger, func(cmd *cobra.Command, args []string) error {
+		RunE: logError(logger, func(cmd *cobra.Command, args []string) error {
 			imageName := cfg.DefaultBuilder
 			if len(args) >= 1 {
 				imageName = args[0]
 			}
 
 			if imageName == "" {
-				SuggestSettingBuilder(logger, inspector)
+				suggestSettingBuilder(logger, inspector)
 				return pack.NewSoftError()
 			}
 
 			builderInfo := writer.SharedBuilderInfo{
 				Name:      imageName,
 				IsDefault: imageName == cfg.DefaultBuilder,
-				Trusted:   IsTrustedBuilder(cfg, imageName),
+				Trusted:   isTrustedBuilder(cfg, imageName),
 			}
 
 			localInfo, localErr := inspector.InspectBuilder(imageName, true, pack.WithDetectionOrderDepth(flags.Depth))
