+++ internal/commands/rebase.go
@@ -23,9 +23,9 @@ func Rebase(logger logging.Logger, cfg config.Config, client PackClient) *cobra.
 		Example: "pack rebase buildpacksio/pack",
 		Long: "Rebase allows you to quickly swap out the underlying OS layers (run image) of an app image generated by `pack build` " +
 			"with a newer version of the run image, without re-building the application.",
-		RunE: LogError(logger, func(cmd *cobra.Command, args []string) error {
+		RunE: logError(logger, func(cmd *cobra.Command, args []string) error {
 			opts.RepoName = args[0]
-			opts.AdditionalMirrors = GetMirrors(cfg)
+			opts.AdditionalMirrors = getMirrors(cfg)
 
 			var err error
 			opts.PullPolicy, err = pubcfg.ParsePullPolicy(policy)
