+++ internal/commands/report.go
@@ -25,7 +25,7 @@ func Report(logger logging.Logger, version string) *cobra.Command {
 		Args:    cobra.NoArgs,
 		Short:   "Display useful information for reporting an issue",
 		Example: "pack report",
-		RunE: LogError(logger, func(cmd *cobra.Command, args []string) error {
+		RunE: logError(logger, func(cmd *cobra.Command, args []string) error {
 			var buf bytes.Buffer
 			err := generateOutput(&buf, version, explicit)
 			if err != nil {
