+++ internal/commands/set_default_registry.go
@@ -17,9 +17,9 @@ func SetDefaultRegistry(logger logging.Logger, cfg config.Config, cfgPath string
 	cmd := &cobra.Command{
 		Use:     "set-default-registry <name>",
 		Args:    cobra.ExactArgs(1),
-		Short:   PrependExperimental("Set default registry"),
+		Short:   prependExperimental("Set default registry"),
 		Example: "pack set-default-registry myregistry",
-		RunE: LogError(logger, func(cmd *cobra.Command, args []string) error {
+		RunE: logError(logger, func(cmd *cobra.Command, args []string) error {
 			registryName = args[0]
 			if !containsRegistry(config.GetRegistries(cfg), registryName) {
 				return errors.Errorf("no registry with the name %s exists", style.Symbol(registryName))
