+++ internal/commands/suggest_stacks.go
@@ -1,27 +1,26 @@
-package stack
+package commands
 
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/buildpacks/pack/internal/commands"
+	"github.com/buildpacks/pack/internal/commands/stack"
 	"github.com/buildpacks/pack/logging"
 )
 
-// Deprecated: Use Suggest instead
+// Deprecated: Use `stack suggest` instead
 func SuggestStacks(logger logging.Logger) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:     "suggest-stacks",
 		Args:    cobra.NoArgs,
 		Short:   "Display list of recommended stacks",
 		Example: "pack suggest-stacks",
-		RunE: commands.LogError(logger, func(cmd *cobra.Command, args []string) error {
+		Run: func(*cobra.Command, []string) {
 			logger.Warn("Command 'pack suggest-stacks' has been deprecated, please use 'pack stack suggest' instead")
-			Suggest(logger)
-			return nil
-		}),
+			stack.Suggest(logger)
+		},
 		Hidden: true,
 	}
 
-	commands.AddHelpFlag(cmd, "suggest-stacks")
+	AddHelpFlag(cmd, "suggest-stacks")
 	return cmd
 }
