+++ internal/commands/suggest_stacks_test.go
@@ -1,4 +1,4 @@
-package stack_test
+package commands_test
 
 import (
 	"bytes"
@@ -10,7 +10,7 @@ import (
 	"github.com/sclevine/spec/report"
 	"github.com/spf13/cobra"
 
-	"github.com/buildpacks/pack/internal/commands/stack"
+	"github.com/buildpacks/pack/internal/commands"
 	"github.com/buildpacks/pack/internal/logging"
 	h "github.com/buildpacks/pack/testhelpers"
 )
@@ -28,7 +28,7 @@ func testSuggestStacksCommand(t *testing.T, when spec.G, it spec.S) {
 	)
 
 	it.Before(func() {
-		command = stack.SuggestStacks(logging.NewLogWithWriters(&outBuf, &outBuf))
+		command = commands.SuggestStacks(logging.NewLogWithWriters(&outBuf, &outBuf))
 	})
 
 	when("#SuggestStacks", func() {
